Removed extra code
